fix(data): stack-pointer instruction assembly needs stack-pointer (#834)
PR #804 removed the `(sp)` operand from stack-pointer-based
instructions. This change was motivated by comparing the assembly with
the RISC-V Sail model.

Unfortunately, the Sail model was incorrect in comparison to the RISC-V
spec, the spec being the current "single source of truth".

Add the stack pointer operand back to these instructions.

Fixes issue #833.