spec: Fix unintentional dead code in branching instructions (#963)
Note: This commit assumes that `jump()` and `jump_halfword()`, which
write to `pc` directly, immediately branch upon being executed. And not
that branching is implicitly deferred to end of the instruction.

Assuming the above, some unditional branch instructions have dead code
following a call to `jump()` or `jump_halfword()`. For instance,
`qc.e.jal` would not write the return address to `ra`.

Signed-off-by: Anton Johansson <anjo@rev.ng>
Co-authored-by: Anton Johansson <anjo@rev.ng>
Co-authored-by: ayosher <181945086+ayosher@users.noreply.github.com>